+++ packages/terra-framework-docs/src/terra-dev-site/doc/notification-dialog/example/NotificationDialogInModalManager.jsx
@@ -76,7 +76,7 @@ const ModalContainer = () => {
 
 const ModalManagerExample = () => (
   <div className={cx('example-wrapper')}>
-    <ModalManager>
+    <ModalManager shouldTrapFocus>
       <ModalContainer />
     </ModalManager>
   </div>