+++ packages/terra-date-time-picker/tests/wdio/date-time-picker-spec.js
@@ -387,19 +387,14 @@ Terra.describeViewports('DateTimePicker', ['tiny', 'large'], () => {
       browser.url('/raw/tests/cerner-terra-framework-docs/date-time-picker/date-time-picker-on-change-raw');
       Terra.hideInputCaret('input[name="terra-date-year-input"]');
       Terra.hideInputCaret('input[name="terra-time-minute-input"]');
-      $('input[name="terra-date-month-input"]').setValue('07');
-      $('input[name="terra-date-day-input"]').setValue('12');
+      $('input[name="terra-date-day-input"]').setValue('01');
       $('input[name="terra-time-hour-input"]').setValue('10');
       $('input[name="terra-time-minute-input"]').setValue('30');
-      $('input[name="terra-date-month-input"]').click();
-      browser.keys('Delete');
-      $('input[name="terra-date-day-input"]').click();
-      browser.keys('Delete');
       $('input[name="terra-date-year-input"]').setValue('2017');
 
       expect($('#iso').getText()).toEqual('');
-      expect($('#input-value').getText()).toEqual('//2017 10:30');
-      expect($('#date-value').getText()).toEqual('//2017');
+      expect($('#input-value').getText()).toEqual('/01/2017 10:30');
+      expect($('#date-value').getText()).toEqual('/01/2017');
       expect($('#time-value').getText()).toEqual('10:30');
       expect($('#ambiguous-date').getText()).toEqual('No');
       expect($('#complete-date').getText()).toEqual('No');
@@ -845,7 +840,7 @@ Terra.describeViewports('DateTimePicker', ['large'], () => {
           expect($('input[name="terra-time-minute-input"]')).toHaveValue((`0${today.minute()}`).slice(-2));
           const hourValue = Number($('input[name="terra-time-hour-input"]').getValue());
           const hourTimeValue = today.hours();
-          const hourInRange = (hourTimeValue === hourValue + 1 || hourTimeValue === hourValue);
+          const hourInRange = (hourTimeValue === hourValue - 1 || hourTimeValue === hourValue);
           expect(hourInRange).toBe(true);
         });
 
@@ -863,7 +858,7 @@ Terra.describeViewports('DateTimePicker', ['large'], () => {
           expect($('input[name="terra-time-minute-input"]')).toHaveValue((`0${today.minute()}`).slice(-2));
           const minuteValue = Number($('input[name="terra-time-minute-input"]').getValue());
           const minuteTimeValue = today.minutes();
-          const minuteInRange = (minuteTimeValue === minuteValue + 1 || minuteTimeValue === minuteValue);
+          const minuteInRange = (minuteTimeValue === minuteValue - 1 || minuteTimeValue === minuteValue);
           expect(minuteInRange).toBe(true);
         });
 
@@ -919,7 +914,7 @@ Terra.describeViewports('DateTimePicker', ['large'], () => {
           expect($('input[name="terra-time-minute-input"]')).toHaveValue((`0${today.minute()}`).slice(-2));
           const minuteValue = Number($('input[name="terra-time-minute-input"]').getValue());
           const minuteTimeValue = today.minutes();
-          const minuteInRange = (minuteTimeValue === minuteValue + 1 || minuteTimeValue === minuteValue);
+          const minuteInRange = (minuteTimeValue === minuteValue - 1 || minuteTimeValue === minuteValue);
           expect(minuteInRange).toBe(true);
         });