+++ packages/terra-notification-dialog/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Added
+  * Added FocusLock to confine focus within the modal popup.
+
 ## 4.39.0 - (January 22, 2024)
 
 * Changed