+++ packages/terra-notification-dialog/CHANGELOG.md
@@ -2,8 +2,8 @@
 
 ## Unreleased
 
-* Added
-  * Added FocusLock to confine focus within the modal popup.
+* Updated
+  * Snapshots updated as per FocusLock implementation.
 
 ## 4.39.0 - (January 22, 2024)