+++ packages/terra-notification-dialog/CHANGELOG.md
@@ -3,7 +3,7 @@
 ## Unreleased
 
 * Updated
-  * Snapshots updated as per FocusLock implementation.
+  * Snapshots updated as per FocusOn implementation.
 
 ## 4.39.0 - (January 22, 2024)