+++ packages/terra-notification-dialog/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 4.40.0 - (February 16, 2024)
+
 * Updated
   * Snapshots updated as per FocusOn implementation.